+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Thu Mar 23 22:46:34 PDT 2017"
+buildTime = "Thu Mar 23 23:42:57 PDT 2017"