+++ CHANGES.md
@@ -60,6 +60,9 @@
   * Add type synonym Goal (for "Symbolic ()"), in order to simplify type signatures
 
   * SBV now properly adds check-sat commands and other directives in debugging output.
+
+  * New examples:
+      - Data.SBV.Examples.Optimization.VM: A server allocation optimization problem
     
 ### Version 5.15, 2017-01-30